Pulled mlleap code into separate class
